Move company selector to own module
