+++ src/modules/auth/forms/index.js
@@ -1,7 +1,5 @@
 const { signInForm, signInSchema, signInApplyErrorState } = require('./sign-in');
-const { selectCompanyForm } = require('./select-company');
 
 exports.signInForm = signInForm;
 exports.signInSchema = signInSchema;
 exports.signInApplyErrorState = signInApplyErrorState;
-exports.selectCompanyForm = selectCompanyForm;
