+++ src/modules/auth/routes.js
@@ -69,30 +69,5 @@ module.exports = {
       description: 'Confirms the user has been signed out of service'
     },
     handler: controller.getSignedOut
-  },
-
-  getSelectCompany: {
-    method: 'GET',
-    path: '/select-company',
-    handler: controller.getSelectCompany,
-    options: {
-
-      description: 'Allows the user to select their company'
-    }
-  },
-
-  postSelectCompany: {
-    method: 'POST',
-    path: '/select-company',
-    handler: controller.postSelectCompany,
-    options: {
-      description: 'Allows the user to select their company',
-      validate: {
-        payload: {
-          company: Joi.number(),
-          csrf_token: Joi.string().guid().required()
-        }
-      }
-    }
   }
 };
