+++ src/modules/routes.js
@@ -21,6 +21,7 @@ const returnsReports = require('./returns-reports/routes');
 const internalSearch = require('./internal-search/routes');
 const waiting = require('./waiting/routes');
 const batchNotifications = require('./batch-notifications/routes');
+const companySelector = require('./company-selector/routes');
 
 module.exports = [
   ...Object.values(addLicencesRoutes),
@@ -45,5 +46,6 @@ module.exports = [
   ...Object.values(returnsReports),
   ...Object.values(internalSearch),
   ...Object.values(waiting),
-  ...Object.values(batchNotifications)
+  ...Object.values(batchNotifications),
+  ...Object.values(companySelector)
 ];
