+++ test/modules/auth/controller.js
@@ -1,16 +1,15 @@
 'use strict';
 
 const { expect } = require('code');
-const Lab = require('lab');
-const lab = exports.lab = Lab.script();
+const { experiment, test, beforeEach } = exports.lab = require('lab').script();
 const sinon = require('sinon');
 const controller = require('../../../src/modules/auth/controller.js');
 
-lab.experiment('getSignout', () => {
+experiment('getSignout', () => {
   let request;
   let h;
 
-  lab.beforeEach(async () => {
+  beforeEach(async () => {
     request = {
       sessionStore: { destroy: sinon.stub().resolves(true) },
       cookieAuth: { clear: sinon.spy() },
@@ -27,24 +26,24 @@ lab.experiment('getSignout', () => {
     await controller.getSignout(request, h);
   });
 
-  lab.test('redirects to the signed-out route', async () => {
+  test('redirects to the signed-out route', async () => {
     const redirectTo = h.redirect.lastCall.args[0];
     expect(redirectTo).to.equal('/signed-out?u=e');
   });
 
-  lab.test('the session is destroyed', async () => {
+  test('the session is destroyed', async () => {
     expect(request.sessionStore.destroy.callCount).to.equal(1);
   });
 
-  lab.test('the auth cookie is cleared', async () => {
+  test('the auth cookie is cleared', async () => {
     expect(request.cookieAuth.clear.callCount).to.equal(1);
   });
 });
 
-lab.experiment('getSignedOut', () => {
+experiment('getSignedOut', () => {
   let h;
 
-  lab.beforeEach(async () => {
+  beforeEach(async () => {
     const request = {
       log: sinon.spy(),
       query: { u: 'e' },
@@ -54,12 +53,12 @@ lab.experiment('getSignedOut', () => {
     await controller.getSignedOut(request, h);
   });
 
-  lab.test('sets the page title', async () => {
+  test('sets the page title', async () => {
     const viewContext = h.view.lastCall.args[1];
     expect(viewContext.pageTitle).to.equal('You are signed out');
   });
 
-  lab.test('sets the surveyType', async () => {
+  test('sets the surveyType', async () => {
     const viewContext = h.view.lastCall.args[1];
     expect(viewContext.surveyType).to.equal('external');
   });
