+++ test/modules/company-selector/forms/select-company.js
@@ -1,7 +1,7 @@
 const { find } = require('lodash');
 const { expect } = require('code');
 const { experiment, test, beforeEach } = exports.lab = require('lab').script();
-const { selectCompanyForm } = require('../../../../src/modules/auth/forms/select-company');
+const { selectCompanyForm } = require('../../../../src/modules/company-selector/forms/select-company');
 
 const getRequest = () => {
   return {
